### 2021-09-23 #3759 @whedon
[PRE REVIEW]:  spectrapepper: A Python toolbox for advanced analysis of spectroscopic data for materials and devices
**Submitting author:** @enricgrau (<a href="http://orcid.org/0000-0002-8357-5824">Enric Grau-Luque</a>)
**Repository:** <a href="https://github.com/spectrapepper/spectrapepper" target ="_blank">https://github.com/spectrapepper/spectrapepper</a>
**Version:** 0.1.0
