### 2021-09-29 #3781 @whedon
[REVIEW]:  spectrapepper: A Python toolbox for advanced analysis of spectroscopic data for materials and devices
**Submitting author:** @enricgrau (<a href="http://orcid.org/0000-0002-8357-5824">Enric Grau-Luque</a>)
**Repository:** <a href="https://github.com/spectrapepper/spectrapepper" target ="_blank">https://github.com/spectrapepper/spectrapepper</a>
**Version:** 0.1.3

- [x] **Contribution and authorship:** Has the submitting author (@enricgrau) made major contributions to the software? Does the full list of paper authors seem appropriate and complete? (See spectrapepper/spectrapepper#3)
