### 2021-09-29 #3781 @rkurchin
☝️ note to @enricgrau to fix that invalid DOI issue at some point (not urgent, but will need to be done eventually)
