+++ paper/paper.bib
@@ -243,7 +243,6 @@ year = {2018}
 @book{VanRossumGuidoDrake2009,
 address = {Scotts Valley, CA},
 author = {{Van Rossum}, Guido and Drake, Fred L.},
-doi = {10.5555/1593511},
 isbn = {1441412697},
 mendeley-groups = {CZGSe Megasample},
 publisher = {CreateSpace},
