### 2021-09-30 #3781 @enricgrau
> ☝️ note to @enricgrau to fix that invalid DOI issue at some point (not urgent, but will need to be done eventually)

Thank you for the note @rkurchin. I've fixed the issue in the .bib file and push the changes into the repository. Any other actions I should take? Thank you again.
