### 2021-09-30 #3781 @danielskatz
@rkurchin - this is how to run the checks - doing that and generating the pdf are separate commands
