### 2021-10-03 #3781 @enricgrau
> Nope, that's fine! I'll just regenerate it here and you can double-check that it all looks good...

Thank you @rkurchin. The proof looks good!
