### 2021-10-12 spectrapepper/spectrapepper#1 @ksunden
Quickstart contains method which does not exist `logo`
In initially installing spectrapepper as part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/3781) I attempted to follow the instructions on the [Quickstart](https://spectrapepper.github.io/spectrapepper/quickstart.html) page, which indicates to use `spep.logo()` to test the installation. No such function exists, so it fails.

### Environment Information

-   spectrapepper version: 0.1.0
-   Python version: 3.10.0
-   Operating System: Debian Linux

### Description

At this stage I was validating the docs/testing install. Just about anything that indicated success would have been okay, but instead I got an `AttributeError`.

### What I Did

```
>>> import spectrapepper as spep
>>> spep.logo()
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
AttributeError: module 'spectrapepper' has no attribute 'logo'
```

I would suggest either removing the line with reference to `logo` or placing some other simple test of the library in its place
