### 2021-10-12 spectrapepper/spectrapepper#3 @ksunden
Joss Authorship
The [JOSS Authorship guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship) allow for a variety of non-code contributions.

In this case, I do not see the author (outside of @enricgrau) represented in the library as I look at it other than in the citations for the submitted paper.

Can you please provide a brief explanation of the contributions to spectrapepper specifically of the following authors listed in the submitted paper:

- Fabien Atlan
- Ignacio Becerril-Romero
- Alejandro Perez-Rodriguez
- Maxim Guc
- Victor Izquierdo-Roca
