### 2021-10-14 2a561a0
Contributing file, codecov, urls to new repository
