+++ docs/_sources/quickstart.rst.txt
@@ -9,7 +9,16 @@ Quickstart
 
         conda install -c conda-forge spectrapepper
 
-3. Test it by printing the logo!::
+
+3. Test it by plotting some data!::
 
         import spectrapepper as spep
-        spep.logo()
+        import matplotlib.pyplot as plt
+
+        data = spep.load_spectras()
+        for i in data[1:]:
+            plt.plot(data[0], i)
+        plt.xlabel('Raman shift ($cm^{-1}$)')
+        plt.ylabel('Intensity (a.u.)')
+        plt.show()
+
