+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,7 +1,7 @@
 contact_links:
     - name: Ask questions
-      url: https://github.com/enricgrau/spectrapepper/discussions/categories/q-a
+      url: https://github.com/spectrapepper/spectrapepper/discussions/categories/q-a
       about: Please ask and answer questions here.
     - name: Ideas
-      url: https://github.com/enricgrau/spectrapepper/discussions/categories/ideas
+      url: https://github.com/spectrapepper/spectrapepper/discussions/categories/ideas
       about: Please share your ideas here.

[…]
