### 2021-11-03 #3781 @enricgrau
@stuartcampbell Thank you for your feedback. We will look into integrating more automated testing progressively and also include some bullet points for this on the contribution guidelines with the following updates.
