### 2021-11-05 d3d7d0d
Added the running median function
Function that computes the running median for n-d arrays.
