### 2021-11-05 cbdd6f2
Added a test for running_median_nd
