### 2021-11-05 df95764
Use prefered matplotlib citation
https://matplotlib.org/stable/citing.html
