### 2021-11-05 #3781 @ksunden
Modulo using the preferred matplotlib citation ([PR submitted](https://github.com/spectrapepper/spectrapepper/pull/8)), I'm ready to sign off on my part.

I do encourage the authors to take a look at implementation of some of their core routines: while performance considerations are not claims made and thus not directly evaluated, there are many instances in the codebase where bare python for loops are used for common calculations such as sums/means. These could be replaced by calls to the appropriate numpy/scipy function (perhaps with an `axis` argument) which would likely be both more performant and easier to read.

Even some of the higher level statistic implemented in this package with pure python are available from scipy, which is already a dependency anyway (this includes the spearman and pearson correlation, though the scipy implementation may not be the desired API for this use case, the numeric computation is the same).

I agree with Stuart that additional testing would be nice, but the most important parts are covered.

I think the documentation would benefit from some additional narrative, as well as including the plots generated by the provided examples.
