### 2021-11-06 #3781 @enricgrau
@ksunden Thank you for your valuable insights.

Indeed, we don't make any performance claims, however, with your advice we will better look at the codebase to replace common calculations by proper numpy/scipy functions. Hopefully for a v0.2.0 we can make performance claims.

As you correctly point out, we do include some functions that are already available on other libraries. This is intentional since we believe that having one single library that encloses all this common practices of the field will make it easier for non-experienced users and also allows us to adapt the input and output format that better suite these particular workflows. The latter also reinforces the need for us to make better descriptions of the functions and examples, as you also mention. We will improve this with following versions and include guidelines for good narrratives in the future.
