### 2021-11-06 #3781 @enricgrau
@rkurchin The citation is fixed, both `.bib` and `.md` files are now updated. Thank you!
