### 2021-11-06 5f85311
editorial tweaks to paper
Signed-off-by: Rachel Kurchin <rkurchin@cmu.edu>
