+++ paper/paper.md
@@ -36,66 +36,64 @@ bibliography: paper.bib
 
 # Statement of need
 
-In the past years, the complexity of novel high-tech materials and devices has increased considerably. This complexity
-is expressed mainly by the increasing amount of components and by a broader range of applications. An example of the
-latter is the last generation of thin film solar cells, which are comprised by several functional micro- and nano-
+In recent years, the complexity of novel high-tech materials and devices has increased considerably. This complexity
+is primarily in the form of increasing numbers of components and broader ranges of applications. An example of the
+latter is the last generation of thin-film solar cells, which comprise several functional micro- and nano-
 layers including back contact, absorber, buffer, and transparent front contact. Most of these layers are complex
-multicomponent compounds (Cu(In,Ga)Se2, Sb2Se3, CdTe, CdS, Zn(O,S), ZnO:Al, etc.) that require a fine-tuning of their
-physicochemical properties to ensure their correct functionality and high peformance [@Chopra2004][@Powalla2018]. In
-this way, the further development of such devices requires advanced characterization and methodologies that allow
+multicomponent compounds (Cu(In,Ga)Se2, Sb2Se3, CdTe, CdS, Zn(O,S), ZnO:Al, etc.) that require fine-tuning of their
+physicochemical properties to ensure functionality and high peformance [@Chopra2004; @Powalla2018]. This embedded complexity means that further development of such devices requires advanced characterization and methodologies that allow
 correlating the physicochemical data of the different layers (chemical composition, structural properties, defect
-concentration, etc.) with the performance of the final devices in a fast, precise and reliable way. In this regard,
+concentration, etc.) with the performance of the final devices in a fast, precise, and reliable way. In this regard,
 non-destructive methodologies based on spectroscopic characterization techniques (Raman, photoluminescence, X-ray
 fluorescence, reflectance, transmittance, etc.) have already been demonstrated to possess a high versatility and
-potential for this type of analyses [@Dimitrievska2019][@Guc2017][@Oliva2016]. These spectroscopy-based methodologies
-can provide deep information that encompasses the complexity of novel materials and devices in a non-destructive way
+potential for this type of analyses [@Dimitrievska2019; @Guc2017; @Oliva2016]. These spectroscopy-based methodologies
+can provide deep information that encompasses the complexity of novel materials and devices in a non-destructive way,
 providing a profound understanding of their properties, failure mechanisms, and possible improvements [@Grau-Luque2021].
 The latest advances in the application of spectroscopic methodologies for complex materials and devices include the
 implementation of combinatorial analysis (CA), artificial intelligence (AI) and machine learning (ML), that have been
 already used in few studies and are slowly becoming more common [@Chen2020]. Furthermore, the widespread use of this
-kind of tools in laboratory environments and in on-line/in-line monitoring at production lines, is foreseen to shorten
+type of tool in both laboratory environments and online/inline monitoring of production lines is predicted to shorten
 development times by a factor of 10, from 10 to 20 years to just a few
-years [@Maine2006][@Mueller2016][@AlanAspuru-Guzik2018][@Correa-Baena2018]. Unfortunately, several barriers for
-researchers to implement CA, AI, and ML have been identified [@Mahmood2021][@Gu2019]. One of them is the proper
+years [@Maine2006; @Mueller2016; @AlanAspuru-Guzik2018; @Correa-Baena2018]. Unfortunately, several barriers for
+researchers to implement CA, AI, and ML remain [@Mahmood2021; @Gu2019]. One of them is the proper
 pre-processing of spectroscopic data that allows not only to emphasize the relevant changes in the spectra, but also to
[…]
