### 2021-11-06 #3781 @rkurchin
I opened https://github.com/spectrapepper/spectrapepper/pull/9 with a few editorial tweaks to the manuscript
