### 2021-11-09 #3781 @enricgrau
> I opened [spectrapepper/spectrapepper#9](https://github.com/spectrapepper/spectrapepper/pull/9) with a few editorial tweaks to the manuscript

@rkurchin Thank you for your suggested tweaks. We believe they do improve the way the manuscript reads and, with just a couple of exceptions, where approved and merged.
