### 2021-11-09 #3781 @rkurchin
Great! The next steps are:
1. Create a new tagged release and comment here with the version number
2. Create an archive of the package (e.g. on Zenodo or figshare) with matching metadata (title, authors, etc.) to this submission and report the associated DOI here
And then we'll be ready to proceed with publication! 🎉
